Refactored naming and Total method for invoice

+++ Xero.Accounting/InvoiceLine.cs
@@ -5,9 +5,9 @@
         public int InvoiceLineId { get; set; }
         public string Description { get; set; }
         public int Quantity { get; set; }
-        public double Cost { get; set; }
+        public decimal Cost { get; set; }
 
-        public double Total()
+        public decimal Total()
         {
             return Cost * Quantity;
         }
